# week-4-workshop/api/openapiv2/notes.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Workshop 3",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Notes",
      "description": "Notes service example",
      "externalDocs": {
        "description": "Find out more about grpc-gateway",
        "url": "https://github.com/grpc-ecosystem/grpc-gateway/blob/main/examples/internal/proto/examplepb/a_bit_of_everything.proto"
      }
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/note": {
      "post": {
        "operationId": "Notes_SaveNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SaveNoteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "info",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1NoteInfo"
            }
          }
        ],
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "x-auth": []
          }
        ]
      }
    },
    "/v1/notes": {
      "get": {
        "operationId": "Notes_ListNotes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListNotesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Notes"
        ]
      }
    },
    "/v1/notes/{noteId}": {
      "get": {
        "operationId": "Notes_GetNoteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetNoteByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "someQueryParam",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "Notes"
        ]
      },
      "delete": {
        "operationId": "Notes_DeleteNoteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteNoteByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Notes"
        ]
      },
      "put": {
        "operationId": "Notes_UpdateNoteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateNoteByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "info",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1NoteInfo"
            }
          }
        ],
        "tags": [
          "Notes"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1DeleteNoteByIDResponse": {
      "type": "object"
    },
    "v1GetNoteByIDResponse": {
      "type": "object"
    },
    "v1ListNotesResponse": {
      "type": "object",
      "properties": {
        "notes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Note"
          }
        }
      }
    },
    "v1Note": {
      "type": "object",
      "properties": {
        "noteId": {
          "type": "string",
          "format": "uint64"
        },
        "info": {
          "$ref": "#/definitions/v1NoteInfo"
        }
      }
    },
    "v1NoteInfo": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "format": "string",
          "example": "Example Title",
          "description": "Note title",
          "title": "title",
          "maxLength": 140,
          "minLength": 3,
          "pattern": "^[a-zA-Z]+$"
        },
        "content": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "Notes full description",
      "title": "NoteInformation",
      "required": [
        "title",
        "content"
      ]
    },
    "v1SaveNoteResponse": {
      "type": "object",
      "properties": {
        "noteId": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1UpdateNoteByIDResponse": {
      "type": "object"
    }
  },
  "securityDefinitions": {
    "x-auth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}